💄 style(Todo): TodoList 스타일 수정
