+++ src/components/todos/TodoEditer.presenter.tsx
@@ -1,5 +1,5 @@
 import { PostTodoRequest } from "../../apis/todos/todosAPI.type";
-import * as S from "../../pages/Todos/Todos.style";
+import * as S from "./TodoEditer.style";
 import { FormPresenter } from "../../types/form/FormPresenter.type";
 
 interface Props extends FormPresenter<PostTodoRequest> {
